Avoid address overflow

+++ src/device_image.h
@@ -45,8 +45,8 @@ public:
 	void download(void* data) const;
 	void fill_zero();
 
-	template <typename T> T* ptr(int y = 0) { return (T*)data + y * step; }
-	template <typename T> const T* ptr(int y = 0) const { return (T*)data + y * step; }
+	template <typename T> T* ptr(int y = 0) { return (T*)data + y * (size_t)step; }
+	template <typename T> const T* ptr(int y = 0) const { return (T*)data + y * (size_t)step; }
 
 	void* data;
 	int rows, cols, step;
